final_project/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
ibm-watson==4.4.1
python-dotenv==0.13.0
pyJWT==1.7.1
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3